Rename properties as suggested
IMPORTANT
---------

The ignoreStale JSON property is now called useStaleReviews. The
semantics of useStaleReviews is the opposite (negated) that of
ignoreStale.

That means that all repos that are configured with ignoreStale set to
true should instead be configured with useStaleReviews set to false or
useStaleReviews absent. And the repositories whose configuration doesn't
contain ignoreStale or whose ignoreStale is set to false should instead
be configured with useStaleReviews set to true.

In any case, if present, ignoreStale should be removed from a repo
configuration as that property is now unrecognised.